+++ src/game/state_checkpoint.rs
@@ -713,6 +713,7 @@ impl Game {
         let mut game = Self {
             format,
             arrived: None,
+            prospective_x: std::cell::Cell::new(None),
             successors: std::collections::HashMap::new(),
             damage_taken_this_turn: checkpoint.damage_taken_this_turn,
             damage_taken_by_group_this_turn: {